refactor: Split ChatMessageToolCallBlock into dedicated components
